refactor: [UF-304] 중복 인터페이스 분리 및 에러 ux ui 개선 #338
