+++ src/api/types/exchange.ts
@@ -8,6 +8,25 @@ export type CarrierType = Carrier;
 export type ExchangeStatus = (typeof EXCHANGE_STATUSES)[number];
 export type MobileDataType = (typeof MOBILE_DATA_TYPES)[number];
 
+// ✅ 구매 상태 enum
+export enum PurchaseStatus {
+  PENDING = 'PENDING',
+  PROCESSING = 'PROCESSING',
+  COMPLETED = 'COMPLETED',
+  FAILED = 'FAILED',
+  CANCELLED = 'CANCELLED',
+}
+
+// ✅ 구매 실패 이유 enum
+export enum PurchaseErrorType {
+  INSUFFICIENT_BALANCE = 'INSUFFICIENT_BALANCE',
+  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
+  PRODUCT_UNAVAILABLE = 'PRODUCT_UNAVAILABLE',
+  NETWORK_ERROR = 'NETWORK_ERROR',
+  SERVER_ERROR = 'SERVER_ERROR',
+  USER_CANCELLED = 'USER_CANCELLED',
+}
+
 export interface ExchangePost {
   postId: number;
   title: string;
@@ -63,16 +82,16 @@ export interface BulkPurchaseParams {
   unitPerZet: number;
 }
 
-// 구매 상태 관련 타입
-export interface PurchaseStatus {
+// ✅ 구매 상태 정보 타입
+export interface PurchaseStatusInfo {
   transactionId: string;
-  status: 'PENDING' | 'PROCESSING' | 'COMPLETED' | 'FAILED' | 'CANCELLED';
+  status: PurchaseStatus;
   createdAt: string;
   completedAt?: string;
   failureReason?: string;
 }
 
-// 구매 이력 관련 타입
+// ✅ 구매 이력 타입
 export interface PurchaseTransaction {
   id: string;
   postId: number;
@@ -82,7 +101,7 @@ export interface PurchaseTransaction {
   dataAmountGB: number;
   carrier: Carrier;
   mobileDataType: MobileDataType;
-  status: PurchaseStatus['status'];
+  status: PurchaseStatus;
   createdAt: string;
   completedAt?: string;
 }
